fix : fix type compatible error

+++ ft_bzero.c
@@ -6,7 +6,7 @@
 /*   By: rjeong <rjeong@student.42seoul.kr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/07/06 19:56:15 by rjeong            #+#    #+#             */
-/*   Updated: 2022/07/06 20:11:24 by rjeong           ###   ########.fr       */
+/*   Updated: 2022/07/09 21:44:02 by rjeong           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	ft_bzero(void *s, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (i < n)
